refactor: create method BlockMerkleRoots::from_transactions
